Refactor main and reorder_page
